+++ apps/smtp/src/modules/event-handlers/default-payloads.ts
@@ -10,6 +10,7 @@ import {
   OrderRefundedWebhookPayloadFragment,
 } from "../../../generated/graphql";
 import {
+  NotifyPayloadAccountChangeEmailConfirmation,
   NotifyPayloadAccountChangeEmailRequest,
   NotifyPayloadAccountConfirmation,
   NotifyPayloadAccountDelete,
@@ -309,7 +310,7 @@ const accountChangeEmailRequestPayload: NotifyPayloadAccountChangeEmailRequest =
   token: "bmt4kc-d6e379b762697f6aa357527af36bb9f6",
   old_email: "test@example.com1",
   new_email: "new.email@example.com1",
-  reset_url:
+  redirect_url:
     "http://example.com/reset?email=user%40example.com&token=bmt4kc-d6e379b762697f6aa357527af36bb9f6",
   channel_slug: "default-channel",
   domain: "demo.saleor.cloud",
@@ -317,7 +318,7 @@ const accountChangeEmailRequestPayload: NotifyPayloadAccountChangeEmailRequest =
   logo_url: "",
 };
 
-const accountChangeEmailConfirmPayload: NotifyPayloadAccountChangeEmailRequest = {
+const accountChangeEmailConfirmPayload: NotifyPayloadAccountChangeEmailConfirmation = {
   user: {
     id: "VXNlcjoxOTY=",
     email: "user@example.com",
@@ -332,9 +333,6 @@ const accountChangeEmailConfirmPayload: NotifyPayloadAccountChangeEmailRequest =
   recipient_email: "user@example.com",
   old_email: "old@example.com",
   new_email: "new@example.com",
-  token: "bmt4kc-d6e379b762697f6aa357527af36bb9f6",
-  reset_url:
-    "http://example.com/reset?email=user%40example.com&token=bmt4kc-d6e379b762697f6aa357527af36bb9f6",
   channel_slug: "default-channel",
   domain: "demo.saleor.cloud",
   site_name: "Saleor e-commerce",